Refactor: Se mueve logica a un archivo util
Se mueve la función `formatMillisToTimeString` a su propio archivo dentro de un paquete `util` para mejorar la estructura del código.

Otros cambios:
- Se reemplaza el título de texto "HARMONIA" por el ícono de la aplicación.
